+++ src/apps/survey/components/main/final/Final.hooks.ts
@@ -3,8 +3,7 @@ import {
   useSurveyListByIdQuery,
   useSurveyMainListQuery,
 } from "@/apps/survey/queries";
-import { useShareLink } from "@/hooks";
-import { useMoveOtherContents } from "@/hooks/useMoveOtherContents";
+import { useMoveOtherContents, useShareLink } from "@/hooks";
 import { useState } from "react";
 import { ChartDataStore } from "./Final.types";
